add other functions to scraper.js module
