+++ scraper.js
@@ -3,9 +3,16 @@
  * @module scraper
  */
 
-const Xray = require('x-ray');
+const Xray   = require('x-ray');
+const parse  = require('url-parse');
+const fs     = require('fs');
+const isHtml = require('is-html');
+const yazl = require('yazl');
+const path = require('path');
+const rimraf = require('rimraf');
+
 const x = Xray();
-const parse = require('url-parse');
+const stat = fs.statSync;
 
 module.exports = {
 
@@ -44,6 +51,62 @@ module.exports = {
           challenge,
           solution
       };
+  },
+
+  /**
+   * Creates a zip file based on the contents of the solutions folder
+   * @requires 'fs'
+   * @requires NPM:yazl
+   * @requires 'path'
+   * @requires 'rimraf'
+   * @param {string} zipFileName The name of the output zipfile
+   * @param {array} pathNames An array of all filenames to be zipped
+   */
+  newArchive(zipFileName, pathNames) {
+      const zipfile = new yazl.ZipFile();
+
+      pathNames.forEach(target => {
+          const joinedTarget = path.join('./solutions/', target);
+          const p = stat(joinedTarget);
+              if (p.isFile()) {
+                  zipfile.addFile(joinedTarget, joinedTarget);
+              }
+      });
+      zipfile.outputStream.pipe(fs.createWriteStream(zipFileName)).on("close", () => {
+          console.log(`
+          I zipped it real good!
+
+          You can collect your zipped archive from the fcc-scraper directory.
+
+          *If you found this useful, star the JacksonBates/fcc-scraper repo on GitHub*
+          Twitter, Instagram and FCC forum: @jacksonbates
+          `);
+          rimraf.sync('./solutions');
+        });
+      zipfile.end();
+  },
+
+  /**
+   * Writes solution files to a solutions folder
+   * @requires 'fs'
+   * @requires NPM:isHtml
+   * @param {object} fileObject Object must contain valid 'challenge' and 'solution' keys
+   * @param {int} count Number of files to be zipped
+   * @param {string} camper The username of the camper to be scraped - taken from rl.question at the bottom of the file
+   * @returns undefined
+   */
+  writeFile(fileObject, count, camper) {
+      const challenge = fileObject.challenge.replace(/%20/g, '-');
+      const solution = fileObject.solution;
+      const dir = './solutions';
+      if (!fs.existsSync(dir)) {
+          fs.mkdirSync(dir);
+      }
+      fs.writeFileSync(`./solutions/${challenge}${isHtml(solution) ? '.html' : '.js'}`, solution);
+      const dirArray = fs.readdirSync(dir);
+      if (dirArray.length === count) {
+          this.newArchive(`${camper}-archive-${+new Date}.zip`, dirArray);
+      }
   },
 
 };
